+++ src/Element/HasUsemap.php
@@ -13,10 +13,10 @@ use UnitEnum;
  *
  * Provides an immutable API for setting the `usemap` attribute on HTML elements.
  *
- * Intended for use in tags and components that require manipulation of the `usemap` attribute.
+ * Intended for use in tag elements that require manipulation of the `usemap` attribute.
  *
  * Key features.
- * - Designed for use in image elements (`<img>`) requiring client-side image map assignment.
+ * - Designed for use in tag elements (`<img>`).
  * - Handles the HTML `usemap` attribute.
  * - Immutable method for setting or overriding the `usemap` attribute.
  * - Supports `string`, `Stringable`, `UnitEnum`, and `null` for flexible image map reference assignment.
@@ -45,8 +45,6 @@ trait HasUsemap
      * unset the attribute.
      *
      * @return static New instance with the updated `usemap` attribute.
-     *
-     * @link https://html.spec.whatwg.org/multipage/embedded-content.html#attr-hyperlink-usemap
      *
      * Usage example:
      * ```php